Fix remaining format specifier and clang-format issues

+++ runtime/core/exec_aten/testing_util/tensor_factory.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include <algorithm>
+#include <cinttypes> // PRId32
 #include <cstdint>
 
 #include <c10/util/irange.h>
@@ -624,14 +625,15 @@ inline void validate_strides(
       if ((strides[i] == strides[j])) {
         ET_CHECK_MSG(
             false,
-            "Stride value and size dont comply at index %d."
-            " strides[%d]: %d, strides[%d] = %d, sizes[%d] = %d, sizes[%d] = %d",
-            static_cast<uint32_t>(i),
-            static_cast<uint32_t>(i),
+            "Stride value and size dont comply at index %zu."
+            " strides[%zu]: %" PRId32 ", strides[%" PRId32 "] = %" PRId32
+            ", sizes[%zu] = %" PRId32 ", sizes[%" PRId32 "] = %" PRId32,
+            i,
+            i,
             strides[i],
             j,
             strides[j],
-            static_cast<uint32_t>(i),
+            i,
             sizes[i],
             j,
             sizes[j]);
